feat: Add User model, migration, and seeder
This commit adds the User model, migration, and seeder files to the project. The User model extends the Eloquent Model class and uses the HasFactory trait. The migration file creates the "users" table with an "id" column and "timestamps" columns. The seeder file is currently empty and will be used to populate the "users" table with data in future commits.