refactor: Update User model with table configuration
This commit updates the User model to include the table configuration. The `$table` property is set to `'users'`, the `$primaryKey` property is set to `'id'`, the `$keyType` property is set to `'int'`, and the `$timestamps` and `$incrementing` properties are set to `true`. This change ensures that the User model is correctly mapped to the 'users' table in the database.