feat: Add columns for address details in addresses table
This commit adds the necessary columns for address details in the "addresses" table. The "street", "city", "province", "country", "postal_code", and "contact_id" columns are added to the table. The "contact_id" column is a foreign key referencing the "id" column in the "contacts" table. This change allows for storing and retrieving address information associated with contacts.